Branch: develop / added clean exit with mongodb shutdown etc. Also added another set of user prefences and taken out hard coded stuff in most places

+++ src/com/pipe42/system/StopProcess.java
@@ -4,13 +4,15 @@ public class StopProcess {
 
     public static Boolean stopProcess(Process p) {
 
-        try {
-            p.destroyForcibly();
-            return true;
-        } catch (Exception e) {
-            e.printStackTrace();
+        if(p != null) {
+            try {
+                p.destroyForcibly();
+                return true;
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
 
-        return null;
+        return false;
     }
 }
